Simplified flow. prod(wut, zzz) and poke(wut) throwing exceptions.
In the previous approach, prod(uint128 Zzz) did go via two inherited functions
before actually getting back to medianizer code. Now the code flow can be followed
more easily as it doesn't use the inherited code.

Additionally, we override the inherited methods so it's not possible
to manipulate the value in the medianizer. Added unit tests for it.